Refactor OAuth providers: extract CachingOAuthProvider base class and split into separate files
- Extract CachingOAuthProvider with shared token caching (double-checked
  locking), retry with backoff (fetch_with_retry), config validation
  (CacheConfig), and token invalidation
- OAuthClientProvider and UamiOAuthProvider now extend CachingOAuthProvider,
  eliminating duplicated caching/retry/validation logic
- Split into separate header/source files per provider:
  OAuthProvider.h, OAuthClientProvider.h, CustomOAuthProvider.h,
  UamiOAuthProvider.h (and corresponding .cpp files)
- Deduplicate set_identity_pool_ids via shared join_with_commas helper
