+++ include/schemaregistry/rest/OAuthProvider.h
@@ -1,8 +1,10 @@
 /**
  * OAuth 2.0 Provider for Schema Registry Client
  *
- * Implements OAuth 2.0 Client Credentials grant (RFC 6749 Section 4.4)
- * with automatic token caching and refresh.
+ * Base classes for OAuth bearer authentication. For concrete providers see:
+ * - OAuthClientProvider.h (Client Credentials grant)
+ * - CustomOAuthProvider.h (user-provided token function)
+ * - UamiOAuthProvider.h (Azure UAMI via IMDS)
  *
  * Authentication Methods (Mutually Exclusive):
  * ClientConfiguration supports three authentication methods:
@@ -18,8 +20,6 @@
 
 #include <chrono>
 #include <functional>
-#include <map>
-#include <memory>
 #include <mutex>
 #include <shared_mutex>
 #include <string>
@@ -100,56 +100,26 @@ struct OAuthToken {
 };
 
 /**
- * OAuth 2.0 Client Credentials provider.
+ * Base class for OAuth providers with automatic token caching and refresh.
  *
- * Implements automatic token fetching using OAuth 2.0 Client Credentials
- * grant flow.
+ * Provides thread-safe token caching using double-checked locking with
+ * shared_mutex. Subclasses implement fetch_token() for the actual token
+ * retrieval (HTTP call + parsing + retry).
  */
-class OAuthClientProvider : public OAuthProvider {
+class CachingOAuthProvider : public OAuthProvider {
  public:
-  /**
-   * Configuration for OAuth Client Credentials flow.
-   */
-  struct Config {
-    // Required OAuth parameters
-    std::string client_id;
-    std::string client_secret;
-    std::string scope;
-    std::string token_endpoint_url;
+  struct CacheConfig {
+    std::string logical_cluster;
+    std::string identity_pool_id;
 
-    // Optional Confluent Cloud parameters
-    // Required for Confluent Cloud
-    std::string logical_cluster;      // Schema Registry logical cluster ID (e.g., "lsrc-12345")
-    std::string identity_pool_id;     // Identity pool ID (e.g., "pool-abcd")
-
-    // Optional retry configuration
     int max_retries{3};
     int retry_base_delay_ms{1000};
     int retry_max_delay_ms{20000};
-
-    // Optional token refresh behavior
-    // Token is refreshed when elapsed time > threshold * total_lifetime
-    // Default: 0.8 (refreshes when 80% of token lifetime has elapsed)
     double token_refresh_threshold{0.8};
-
-    // Optional HTTP timeout
     int http_timeout_seconds{30};
 
-    /**
-     * Validate configuration has all required fields.
-     *
-     * @throws std::invalid_argument if configuration is invalid
-     */
     void validate() const;
   };
-
-  /**
-   * Construct OAuth provider with configuration.
-   *
-   * @param config OAuth configuration (validated on construction)
-   * @throws std::invalid_argument if configuration is invalid
-   */
-  explicit OAuthClientProvider(const Config& config);
 
   BearerFields get_bearer_fields() override;
 
@@ -158,117 +128,55 @@ class OAuthClientProvider : public OAuthProvider {
    */
   void invalidate_token();
 
- private:
+ protected:
+  explicit CachingOAuthProvider(const CacheConfig& cache_config);
+
   /**
-   * Fetch new token from OAuth provider using Client Credentials grant.
+   * Fetch a new token from the token endpoint.
    *
-   * Performs exponential backoff with jitter on failures.
+   * Subclasses implement the HTTP request and response parsing.
    *
-   * @return Newly fetched OAuth token
+   * @return Newly fetched OAuthToken
    * @throws std::runtime_error if unable to fetch token after retries
    */
-  OAuthToken fetch_token();
-
-  const Config config_;
-
-  mutable std::shared_mutex mutex_;  // Shared mutex for thread-safe token access
-  OAuthToken token_;
-};
-
-/**
- * Custom OAuth provider that delegates token fetching to a user-provided function.
- *
- * This allows users to implement custom token fetching logic without implementing
- * a full provider class.
- *
- * Example usage:
- * @code
- * auto provider = std::make_shared<CustomOAuthProvider>(
- *     []() { return fetch_token_from_my_idp(); },
- *     "lsrc-12345",
- *     "pool-67890"
- * );
- * @endcode
- *
- * Note: The function is called on every get_bearer_fields() call.
- * Users should implement their own caching and thread safety if needed.
- */
-class CustomOAuthProvider : public OAuthProvider {
- public:
-  /**
-   * Function type that returns an access token string.
-   * The function should handle all token fetching logic including:
-   * - Fetching from custom OAuth endpoints
-   * - Caching (if desired)
-   * - Thread safety (if needed)
-   * - Error handling
-   *
-   * @return Access token string
-   * @throws std::runtime_error if unable to fetch token
-   */
-  using TokenFetchFunction = std::function<std::string()>;
+  virtual OAuthToken fetch_token() = 0;
 
   /**
-   * Construct custom OAuth provider with user-provided token fetch function.
-   *
-   * @param fetch_fn Function that returns access token (required)
-   * @param logical_cluster Schema Registry logical cluster ID (optional, required for Confluent Cloud)
-   * @param identity_pool_id Identity pool ID (optional, required for Confluent Cloud)
-   * @throws std::invalid_argument if fetch_fn is empty
+   * Token response from an HTTP request, used by fetch_with_retry.
    */
-  CustomOAuthProvider(TokenFetchFunction fetch_fn,
-                     std::string logical_cluster = "",
-                     std::string identity_pool_id = "");
-
-  BearerFields get_bearer_fields() override;
-
- private:
-  const TokenFetchFunction fetch_fn_;
-  const std::string logical_cluster_;
-  const std::string identity_pool_id_;
-};
-
-/**
- * Azure User-Assigned Managed Identity (UAMI) OAuth provider.
- *
- * Fetches tokens from the Azure Instance Metadata Service (IMDS) using
- * a managed identity's client ID. Includes automatic token caching and
- * refresh with the same double-checked locking pattern as OAuthClientProvider.
- */
-class UamiOAuthProvider : public OAuthProvider {
- public:
-  struct Config {
-    // Azure IMDS base URL (default: standard Azure IMDS endpoint)
-    std::string uami_url = "http://169.254.169.254/metadata/identity/oauth2/token";
-
-    // Required: query string appended to uami_url (contains resource, client_id, api-version, etc.)
-    std::string uami_endpoint_query;
-
-    // Confluent Cloud parameters
-    std::string logical_cluster;
-    std::string identity_pool_id;
-
-    // Retry configuration
-    int max_retries{3};
-    int retry_base_delay_ms{1000};
-    int retry_max_delay_ms{20000};
-    double token_refresh_threshold{0.8};
-    int http_timeout_seconds{30};
-
-    void validate() const;
-    void set_identity_pool_ids(const std::vector<std::string>& pool_ids);
+  struct TokenResponse {
+    int status_code{0};
+    std::string text;
   };
 
-  explicit UamiOAuthProvider(const Config& config);
+  /**
+   * Execute an HTTP request with retry and exponential backoff.
+   *
+   * Retries on network errors and retriable HTTP status codes.
+   * Returns the successful response body for the caller to parse.
+   *
+   * @param provider_name Name for error messages (e.g., "OAuth", "UAMI")
+   * @param make_request Function that performs the HTTP request
+   * @return Response body text on success (2xx)
+   * @throws std::runtime_error if unable to get a successful response after retries
+   */
+  std::string fetch_with_retry(
+      const std::string& provider_name,
+      const std::function<TokenResponse()>& make_request);
 
-  BearerFields get_bearer_fields() override;
+  /**
+   * Parse a standard OAuth token response (access_token string, expires_in int).
+   *
+   * @param json_text Raw JSON response body
+   * @param provider_name Name for error messages
+   * @return Parsed OAuthToken
+   */
+  static OAuthToken parse_standard_token_response(
+      const std::string& json_text, const std::string& provider_name);
 
-  void invalidate_token();
+  const CacheConfig cache_config_;
 
  private:
-  OAuthToken fetch_token();
-
-  const Config config_;
   mutable std::shared_mutex mutex_;
   OAuthToken token_;
 };
